Add React Js
This is a single-page application website. I used for the project HTML CSS JAVASCRIPT BOOTSTRAP and REACT JS. This is also a responsive website. For responsiveness, I used a Media query. This is a SPA is single page application. The way we build our app is no different than all the apps we've been building so far. We will have a main parent component. Each of our app's individual "pages" will be separate components that feed into the main component. The magic React Router brings to the table is basically choosing which components to show and which to hide